+++ app/src/main/java/org/oxycblt/auxio/playback/PlaybackPanelFragment.kt
@@ -42,7 +42,7 @@ import org.oxycblt.auxio.playback.state.RepeatMode
 import org.oxycblt.auxio.playback.ui.StyledSeekBar
 import org.oxycblt.auxio.ui.ViewBindingFragment
 import org.oxycblt.auxio.util.collectImmediately
-import org.oxycblt.auxio.util.shareSong
+import org.oxycblt.auxio.util.share
 import org.oxycblt.auxio.util.showToast
 import org.oxycblt.auxio.util.systemBarInsetsCompat
 
@@ -174,7 +174,7 @@ class PlaybackPanelFragment :
                 true
             }
             R.id.action_share -> {
-                playbackModel.song.value?.let { requireContext().shareSong(it) }
+                playbackModel.song.value?.let { requireContext().share(it) }
                 true
             }
             else -> false
